Wstępna obsługa repozytoriów maven i plików .jar

+++ src/main/java/org/agiso/tempel/engine/VelocityFileEngine.java
@@ -7,15 +7,14 @@
 package org.agiso.tempel.engine;
 
 import java.io.File;
-import java.io.FileReader;
 import java.io.FileWriter;
-import java.io.IOException;
+import java.io.InputStreamReader;
 import java.io.Reader;
 import java.io.Writer;
 import java.util.Map;
-import java.util.zip.ZipFile;
 
 import org.agiso.tempel.core.ITemplateSource;
+import org.agiso.tempel.core.ITemplateSourceEntry;
 import org.apache.velocity.VelocityContext;
 
 /**
@@ -28,32 +27,53 @@ public class VelocityFileEngine extends AbstractVelocityEngine {
 	public void run(ITemplateSource templateSource, Map<String, Object> params, String target) {
 		// Wyznaczanie ścieżki zasobu docelowego i sprawdzanie jego istnienia:
 		if(!templateSource.exists()) {
-			throw new RuntimeException("Zasób " + templateSource.getResource() + " nie istnieje");
+			throw new RuntimeException("Zasób "
+					+ templateSource.getTemplate() + "/"+ templateSource.getResource()
+					+ " nie istnieje"
+			);
 		}
 
 		// Tworzenie kontekstu Velocity wspólnego w całym procesie obsługi:
 		VelocityContext context = createVelocityContext(params);
 
 		try {
-			processVelocityResource(templateSource.getEntry(), context, target);
-		} catch(IOException e) {
+			processVelocityResource(templateSource, context, target);
+		} catch(Exception e) {
 			throw new RuntimeException(e);
 		}
 	}
 
 //	--------------------------------------------------------------------------
-	protected void processVelocityResource(File resource, VelocityContext context, String target) throws IOException {
-		processVelocityFile(resource, context, target);
+	/**
+	 * Szablon musi być pojedynczym plikiem. Jego przetwarzanie polega na jednorazowym
+	 * wywołaniu metody {@link #processVelocityFile(ITemplateSourceEntry, VelocityContext,
+	 * String)} w celu utworzenia pojedynczego zasobu.
+	 */
+	protected void processVelocityResource(ITemplateSource source, VelocityContext context, String target) throws Exception {
+		if(!source.isFile()) {
+			throw new RuntimeException("Zasób "
+					+ source.getTemplate() + "/"+ source.getResource()
+					+ " nie jest plikiem"
+			);
+		}
+		processVelocityFile(source.getEntry(source.getResource()), context, target);
 	}
 
-	protected final void processVelocityFile(File resource, VelocityContext context, String target) throws IOException {
+	protected void processVelocityFile(ITemplateSourceEntry entry, VelocityContext context, String target) throws Exception {
+		if(!entry.isFile()) {
+			throw new RuntimeException("Element "
+					+ entry.getTemplate() + "/"+ entry.getName()
+					+ " nie jest plikiem"
+			);
+		}
+
 		Writer writer = null;
 		Reader reader = null;
 		try {
-			reader = new FileReader(resource);
+			reader = new InputStreamReader(entry.getInputStream());
 			writer = new FileWriter(new File(target));
 
-			doVelocityTemplateMerge(resource.getPath(), reader, context, writer);
+			doVelocityTemplateMerge(entry.getName(), reader, context, writer);
 		} finally {
 			if(reader != null) {
 				try {
